+++ frontend/src/main.tsx
@@ -1,18 +1,22 @@
-// TODO: restore auth wrappers when re-enabling login
-// import './auth/captureGoogleHash'
-// import { MsalProvider } from '@azure/msal-react'
-// import { msalInstance } from './auth/msalInstance'
-// import { validateEnv } from './validateEnv'
+import './auth/captureGoogleHash';
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { MsalProvider } from '@azure/msal-react';
+import { GoogleOAuthProvider } from '@react-oauth/google';
+import { msalInstance } from './auth/msalInstance';
+import { googleClientId } from './auth/googleConfig';
+import { validateEnv } from './validateEnv';
+import './index.css';
+import App from './app/App';
 
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './app/App'
-
-// validateEnv();
+validateEnv();
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <GoogleOAuthProvider clientId={googleClientId}>
+      <MsalProvider instance={msalInstance}>
+        <App />
+      </MsalProvider>
+    </GoogleOAuthProvider>
   </StrictMode>,
-)
+);